mensa_bot/requirements.txt
python-dotenv==0.3.0
python-telegram-bot==3.2.0
PyYAML==3.11
.
